Fix default page header handling
Until now, the default content of page header was included
in the baseof template. In child templates for which the
header was supposed to be empty, it was overriden by defining
the content of its block as a whitespace character. This character
was, however, still visible and forcing the header to include
a blank line.

This commit fixes that by moving the common page header content
to a new partial template: page_header_content and including this
partial template in page header content definitions in templates
that require it.